requirements.txt
fpdf2==2.7.4
python-dotenv